chore(dal-test): Reorganise the test schemas to install only in dal-test
This stops the dal crate knowing anything about the test specific schemas
