+++ lib/dal-test/src/schemas/mod.rs
@@ -1,0 +1,7 @@
+mod test_exclusive_schema_bethesda_secret;
+mod test_exclusive_schema_fallout;
+mod test_exclusive_schema_starfield;
+
+pub use test_exclusive_schema_bethesda_secret::migrate_test_exclusive_schema_bethesda_secret;
+pub use test_exclusive_schema_fallout::migrate_test_exclusive_schema_fallout;
+pub use test_exclusive_schema_starfield::migrate_test_exclusive_schema_starfield;
