+++ lib/dal/src/builtins/schema.rs
@@ -1,147 +1,9 @@
 use si_pkg::SiPkg;
-use strum::{AsRefStr, Display, EnumIter, EnumString};
 use telemetry::prelude::*;
 
 use crate::installed_pkg::InstalledPkg;
 use crate::pkg::{import_pkg_from_pkg, ImportOptions};
 use crate::{BuiltinsError, BuiltinsResult, DalContext};
-
-mod test_exclusive_schema_bethesda_secret;
-mod test_exclusive_schema_fallout;
-mod test_exclusive_schema_starfield;
-
-pub use test_exclusive_schema_bethesda_secret::migrate_test_exclusive_schema_bethesda_secret;
-pub use test_exclusive_schema_fallout::migrate_test_exclusive_schema_fallout;
-pub use test_exclusive_schema_starfield::migrate_test_exclusive_schema_starfield;
-
-// /// Migrate [`Schemas`](crate::Schema) for production use.
-// pub async fn migrate_for_production(ctx: &DalContext) -> BuiltinsResult<()> {
-//     info!("migrating schemas");
-//
-//     migrate_pkg(ctx, super::SI_AWS_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_AWS_EC2_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_DOCKER_IMAGE_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_COREOS_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_GENERIC_FRAME_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_AWS_IAM_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_AWS_ECS_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_AWS_CLOUDWATCH_PKG, None).await?;
-//     migrate_pkg(ctx, super::SI_AWS_LB_TARGET_GROUP_PKG, None).await?;
-//
-//     Ok(())
-// }
-
-#[remain::sorted]
-#[derive(Debug, Copy, Clone, AsRefStr, Display, EnumIter, EnumString, Eq, PartialEq)]
-pub enum BuiltinSchema {
-    BethesdaSecret,
-    Fallout,
-    Starfield,
-}
-
-impl BuiltinSchema {
-    pub fn real_schema_name(&self) -> &'static str {
-        match self {
-            BuiltinSchema::BethesdaSecret => "bethesda-secret",
-            BuiltinSchema::Fallout => "fallout",
-            BuiltinSchema::Starfield => "starfield",
-        }
-    }
-}
-
-// pub async fn migrate_schema(
-//     ctx: &DalContext,
-//     schema: BuiltinSchema,
-//     driver: &MigrationDriver,
-// ) -> BuiltinsResult<()> {
-//     match schema {
-//         BuiltinSchema::Fallout => {
-//             driver.migrate_test_exclusive_fallout(ctx).await?;
-//         }
-//         BuiltinSchema::Starfield => {
-//             driver.migrate_test_exclusive_starfield(ctx).await?;
-//         }
-//     }
-//
-//     Ok(())
-// }
-
-// /// Migrate [`Schemas`](crate::Schema) for use in tests.
-// pub async fn migrate_for_tests(
-//     ctx: &DalContext,
-//     selected_test_builtin_schemas: SelectedTestBuiltinSchemas,
-// ) -> BuiltinsResult<()> {
-//     // Determine what to migrate based on the selected test builtin schemas provided.
-//     let (migrate_all, migrate_test_exclusive, specific_builtin_schemas) =
-//         match selected_test_builtin_schemas {
-//             SelectedTestBuiltinSchemas::All => {
-//                 info!("migrating schemas for tests");
-//                 (true, false, HashSet::new())
-//             }
-//             SelectedTestBuiltinSchemas::None => {
-//                 info!("skipping migrating schemas for tests");
-//                 return Ok(());
-//             }
-//             SelectedTestBuiltinSchemas::Some(provided_set) => {
-//                 info!("migrating schemas for tests based on a provided set of names");
-//                 debug!("provided set of builtin schemas: {:?}", &provided_set);
-//                 (false, false, provided_set)
-//             }
-//             SelectedTestBuiltinSchemas::Test => {
-//                 info!("migrating test-exclusive schemas solely");
-//                 (false, true, HashSet::new())
-//             }
-//         };
-
-//     // Once we know what to migrate, create the driver.
-//     let driver = MigrationDriver::new(ctx).await?;
-//     ctx.blocking_commit().await?;
-
-//     if migrate_all {
-//         migrate_pkg(ctx, super::SI_AWS_PKG, None).await?;
-//         migrate_pkg(ctx, super::SI_AWS_EC2_PKG, None).await?;
-//         migrate_pkg(ctx, super::SI_COREOS_PKG, None).await?;
-//         migrate_pkg(ctx, super::SI_DOCKER_IMAGE_PKG, None).await?;
-//         migrate_pkg(ctx, super::SI_GENERIC_FRAME_PKG, None).await?;
-//         migrate_pkg(ctx, super::SI_AWS_LB_TARGET_GROUP_PKG, None).await?;
-//         for test_schema in [BuiltinSchema::Starfield, BuiltinSchema::Fallout] {
-//             migrate_schema(ctx, test_schema, &driver).await?;
-//             ctx.blocking_commit().await?;
-//         }
-//     } else if migrate_test_exclusive {
-//         // We migrate generic frame to get "si:resourceToPayloadValue" cheaply. This function
-//         // should be converted to an intrinsic (or removed?)
-//         migrate_pkg(ctx, super::SI_GENERIC_FRAME_PKG, None).await?;
-//         for test_schema in [BuiltinSchema::Starfield, BuiltinSchema::Fallout] {
-//             migrate_schema(ctx, test_schema, &driver).await?;
-//             ctx.blocking_commit().await?;
-//         }
-//     } else {
-//         let schemas: Vec<String> = specific_builtin_schemas
-//             .iter()
-//             .map(|s| s.to_owned())
-//             .collect();
-//         migrate_pkg(ctx, super::SI_AWS_PKG, Some(schemas.to_owned())).await?;
-//         migrate_pkg(ctx, super::SI_AWS_EC2_PKG, Some(schemas.to_owned())).await?;
-//         migrate_pkg(ctx, super::SI_COREOS_PKG, Some(schemas.to_owned())).await?;
-//         migrate_pkg(ctx, super::SI_DOCKER_IMAGE_PKG, Some(schemas.to_owned())).await?;
-//         migrate_pkg(ctx, super::SI_GENERIC_FRAME_PKG, Some(schemas.to_owned())).await?;
-//         migrate_pkg(
-//             ctx,
-//             super::SI_AWS_LB_TARGET_GROUP_PKG,
-//             Some(schemas.to_owned()),
-//         )
-//         .await?;
-//         for test_schema in [BuiltinSchema::Starfield, BuiltinSchema::Fallout] {
-//             if specific_builtin_schemas.contains(test_schema.real_schema_name()) {
-//                 migrate_schema(ctx, test_schema, &driver).await?;
-//                 ctx.blocking_commit().await?;
-//             }
-//         }
-//     }
-
-//     Ok(())
-// }
 
 pub async fn migrate_pkg(
     ctx: &DalContext,
